+++ docs/index.html
@@ -13,9 +13,6 @@
     <link rel="shortcut icon" type="image/x-icon" href="img/favicon.ico">
     <script src="https://code.jquery.com/jquery-3.3.1.min.js"
         integrity="sha256-FgpCb/KJQlLNfOu91ta32o/NMZxltwRo8QtmkMRdAu8=" crossorigin="anonymous"></script>
-    <script src="js/jquery.scrollify.js"></script>
-    <script src="js/scroll.js"></script>
-    <script src="js/nav.js"></script>
     <script src="./bundle.js"></script>
     <meta name="google-site-verification" content="y5QwA76yrPZYH9bjFGbaq0F6C0Xp35ydYY2LK6nmjlY" />
 </head>
