+++ package.json
@@ -24,5 +24,8 @@
     "webpack": "^4.33.0",
     "webpack-cli": "^3.3.3",
     "webpack-dev-server": "^3.7.1"
+  },
+  "dependencies": {
+    "jquery-scrollify": "^1.0.20"
   }
 }
