+++ src/js/index.js
@@ -1,1 +1,2 @@
-console.log( 'sample' );
+import './scroll';
+import './nav';
